Pagination ready.

+++ services/ShoppingCartService.class.php
@@ -12,7 +12,12 @@ class ShoppingCartService
     public function get_cart($user_id, $session_id)
     {
         $shopping_cart = ShoppingCart::get_by($user_id, $session_id);
-        $shopping_cart->get_items();
+
+        if ($shopping_cart)
+        {
+            $shopping_cart->get_items();
+        }
+
         return $shopping_cart;
     }
 
@@ -20,14 +25,17 @@ class ShoppingCartService
     {
         $shopping_cart = ShoppingCart::get_by($user_id, $session_id);
 
-        foreach ($shopping_cart->get_items() as $cart_item)
+        if ($shopping_cart)
         {
-            $cart_item->product->increase_quantity_in_stock_by($cart_item->product_quantity);
-            $cart_item->product->update();
-            $cart_item->delete();
-        }
+            foreach ($shopping_cart->get_items() as $cart_item)
+            {
+                $cart_item->product->increase_quantity_in_stock_by($cart_item->product_quantity);
+                $cart_item->product->update();
+                $cart_item->delete();
+            }
 
-        $shopping_cart->calculate_total();
-        $shopping_cart->update();
+            $shopping_cart->calculate_total();
+            $shopping_cart->update();
+        }
     }
 }
